07_proyectos/principiante/proyecto_01_etl_simple/requirements.txt
# Dependencias para el Pipeline ETL Simple
pandas>=2.0.0
python-dotenv>=1.0.0
pyarrow>=12.0.0  # Para soporte de Parquet
